Moved Search Bar to Pets page
