--- src/pages/search.js
@@ -1,12 +1,0 @@
-import React from "react";
-import SearchBar from "../components/SearchBar/SearchBar";
- 
-const Search = () => {
-    return (
-        <div>
-            <SearchBar />
-        </div>
-    );
-};
- 
-export default Search;
